Completed Money and UserBalance classes

+++ src/main/java/com/techelevator/currency/Depositable.java
@@ -1,5 +1,6 @@
 package com.techelevator.currency;
+import java.math.BigDecimal;
 
 public interface Depositable {
-    public void deposit(double deposit);
+    public void deposit(BigDecimal deposit);
 }
